Print the bytes with %u, and let the CI show what gcc says
The first run with -Werror failed seven seconds into the build, about
where core.o is made, and core.c holds a printf format gcc on mingw is
known to refuse. vangopix_size_text printed the bytes below a kilobyte
with %llu, and gcc there checks a format against the rules of the
Microsoft runtime, where ll is not a length - a warning every build
before had printed and gone on from, and one clang never gives, which
is why the suite passed at home. Below the first unit the number is
under 1024, so it is printed %u: the one spelling every C runtime
agrees on.

The log that would have said so for certain needs a login, so two
changes to the workflow. A problem matcher (.github/gcc.json) turns
gcc's own warning and error lines into annotations on the commit and
the pull request, pinned to the file and line - where a warning is read
the day it arrives, and readable without signing in. And make -k keeps
going past the first object that fails, so one run names every file
with something to fix instead of one file per push.

Checked with make test WERROR=1 on clang: no warnings, every check
passing. Whether gcc has anything more to say is what this run is for.

+++ src/core.c
@@ -597,7 +597,13 @@ void vangopix_size_text (Uint64 bytes, char *dst, size_t cap)
 		u++;
 	}
 
-	if (u == 0) SDL_snprintf(dst, cap, "%llu B", (unsigned long long)bytes);
+	/*
+	 * %u, NOT %llu, for a Uint64: below the first unit the number is under 1024 and fits any
+	 * unsigned, and %u is the one spelling every C runtime agrees on. gcc on mingw checks a
+	 * printf format against the rules of the Microsoft runtime, and ll is not in them - a
+	 * warning at home, and a failed build under the CI's -Werror.
+	 */
+	if (u == 0) SDL_snprintf(dst, cap, "%u B", (unsigned)bytes);
 	else        SDL_snprintf(dst, cap, "%.1f %s", v, UNIT[u]);
 }
